requirements.txt
GitPython
tinydb
textual
psutil
